Revert "Project refactoring"
This reverts commit e71ec8fd8c5b0749e8802d0e6df3b2362ab75219.
I'm re-refactoring the source code hierarchy based on comments I've
heard about my weird golang source code tree hierarchy choices someone
out there made.
