+++ lib/local/known_hosts.go
@@ -4,8 +4,8 @@ import (
 	"bufio"
 	"errors"
 	"fmt"
-	"github.com/bgartzi/uhhm/lib/host"
-	"github.com/bgartzi/uhhm/lib/utils"
+	"github.com/bgartzi/uhhmm/lib/host"
+	"github.com/bgartzi/uhhmm/lib/utils"
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/crypto/ssh/knownhosts"
 	"os"
